requirements.txt
asgiref==3.8.1
certifi==2024.8.30
Django==4.2
django-cors-headers==4.6.0
django-jazzmin==3.0.0
django-ranged-response==0.2.0
django-simple-captcha==0.6.0
django-unfold==0.43.0
djangorestframework==3.15.2
pillow==10.4.0
sqlparse==0.5.1
